+++ src/main/java/com/YEP4510/YEP4510/Repositories/OficialRepository.java
@@ -4,5 +4,5 @@ import com.YEP4510.YEP4510.Models.Oficial;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface OficialRepository extends CrudRepository<Oficial, Integer>, PagingAndSortingRepository<Oficial, Integer> {
+public interface OficialRepository extends CrudRepository<Oficial, Long>, PagingAndSortingRepository<Oficial, Long> {
 }
