+++ src/main/java/com/YEP4510/YEP4510/Repositories/ResponsavelRepository.java
@@ -4,5 +4,5 @@ import com.YEP4510.YEP4510.Models.Responsavel;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface ResponsavelRepository extends CrudRepository<Responsavel, Integer>, PagingAndSortingRepository<Responsavel, Integer> {
+public interface ResponsavelRepository extends CrudRepository<Responsavel, Long>, PagingAndSortingRepository<Responsavel, Long> {
 }
