+++ src/main/java/com/YEP4510/YEP4510/Repositories/UsuarioRepository.java
@@ -4,5 +4,5 @@ import com.YEP4510.YEP4510.Models.Usuario;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface UsuarioRepository extends CrudRepository<Usuario, Integer>, PagingAndSortingRepository<Usuario, Integer> {
+public interface UsuarioRepository extends CrudRepository<Usuario, Long>, PagingAndSortingRepository<Usuario, Long> {
 }
